Resolved screen issue and formatting

+++ templates/bootstrap.tpl
@@ -7,11 +7,23 @@ set -x
 
 ## Global Vars
 JAR_NAME="server.jar"
+SCREEN_SES="mc_server"
+SCREEN_CMD="java -Xms${server_min_ram} -Xmx${server_max_ram} -jar ${mc_home_folder}/$JAR_NAME nogui"
 
 prepare_ssd() {
     mkdir ${mc_home_folder}
-    mkfs -t ext4 /dev/sdb
+    if [ "$(blkid -o value --match-tag TYPE /dev/sdb)" != "ext4" ]; then
+        mkfs -t ext4 /dev/sdb
+    fi
     mount -t ext4 /dev/sdb ${mc_home_folder}
+    
+    if [ ! -d "${mc_script_location}" ]; then
+        mkdir ${mc_script_location}
+
+        configure_backups
+        configure_restart
+        configure_restore_backup
+    fi
 }
 
 install_pre_req() {
@@ -20,70 +32,65 @@ install_pre_req() {
 
     apt-get update
     apt-get install -y gsutil
-    apt-get install -y default-jre-headless zip screen google-cloud-sdk
+    apt-get install -y default-jre-headless zip screen less google-cloud-sdk
 }
 
 setup_mc_server() {
     cd ${mc_home_folder}
-    wget -O $JAR_NAME ${mc_server_download_link}
+    if [ ! -f "${mc_home_folder}/$JAR_NAME" ]; then
+        wget -O ${mc_home_folder}/$JAR_NAME ${mc_server_download_link}
 
-    set +e
-    java -Xms${server_min_ram} -Xmx${server_max_ram} -jar $JAR_NAME nogui
-    set -e
+        set +e
+        $SCREEN_CMD
+        set -e
 
-    sed -i 's/eula=false/eula=true/g' eula.txt
-    touch eula.txt
+        sed -i 's/eula=false/eula=true/g' ${mc_home_folder}/eula.txt
+    fi
 
-    screen -dm -S mcs /bin/bash -c 'java -Xms${server_min_ram} -Xmx${server_max_ram} -jar $JAR_NAME nogui; exec /bin/bash'
+    command="$SCREEN_CMD" screen -S $SCREEN_SES -d -m bash -c '$command; exec bash'
 }
 
 configure_backups() {
-    cd ${mc_home_folder}
-
-    cat << EOF > ./backup.sh
+    cat << EOF > ${mc_script_location}/backup.sh
 #!/bin/bash
-screen -r mcs -X stuff '/save-all\n/save-off\n'
+cd ${mc_home_folder}
+screen -r $SCREEN_SES -X stuff '/save-all\n/save-off\n'
 /usr/bin/zip -r -q backup.zip ${mc_home_folder}/world
 /usr/bin/gsutil cp -R ${mc_home_folder}/backup.zip gs://${backup_bucket}/\$(date "+%Y%m%d-%H%M%S")-world
 rm -f backup.zip
-screen -r mcs -X stuff '/save-on\n'
+screen -r $SCREEN_SES -X stuff '/save-on\n'
 EOF
 
-    chmod 755 ./backup.sh
+    chmod 755 ${mc_script_location}/backup.sh
 }
 
 configure_restart() {
-    cd ${mc_home_folder}
-
-    cat << EOF > ./restart.sh
+    cat << EOF > ${mc_script_location}/restart.sh
 #!/bin/bash
 mount -t ext4 /dev/sdb ${mc_home_folder}
-screen -dm -S mcs /bin/bash -c 'java -Xms${server_min_ram} -Xmx${server_max_ram} -jar $JAR_NAME nogui; exec /bin/bash'
+cd ${mc_home_folder}
+command="$SCREEN_CMD" screen -S $SCREEN_SES -d -m bash -c '\$command; exec bash'
 EOF
 
-    chmod 755 ./restart.sh
+    chmod 755 ${mc_script_location}/restart.sh
 }
 
 configure_restore_backup() {
-     cd ${mc_home_folder}
-
-     cat << EOF > ./restore_backup.sh
+     cat << EOF > ${mc_script_location}/restore_backup.sh
 #!/bin/bash
-screen -r -X stuff '/stop\\n'
+cd ${mc_home_folder}
+screen -S $SCREEN_SES -r -X stuff '/stop\n'
 mv world/ world_old/
 /usr/bin/gsutil cp gs://${backup_bucket}/\$1 backup.zip
 mkdir world
 /usr/bin/unzip -q backup.zip -d world/
-sleep 30
-screen -dm -S mcs /bin/bash -c 'java -Xms${server_min_ram} -Xmx${server_max_ram} -jar $JAR_NAME nogui; exec /bin/bash'
+command="$SCREEN_CMD" screen -S $SCREEN_SES -d -m bash -c '\$command; exec bash'
 rm -f backup.zip
 EOF
+    chmod 755 ${mc_script_location}/restore_backup.sh
 }
 
-## Main Execution
+## MAIN
 install_pre_req
 prepare_ssd
 setup_mc_server
-configure_backups
-configure_restart
-configure_restore_backup
